refactor(task-graph): extract RunScheduler from TaskGraphRunner
Moves the run-loop coordinator (pushStatusFromNodeToEdges,
propagateDisabledStatus, handleProgress, plus arm/clear timeout helpers and
the runGraph for-await loop body) into a new internal class. The facade keeps
runGraph as the lifecycle orchestrator + the terminal-state precedence
(timeout > error > abort > complete), but the parallel-task scheduling and
status-push concerns live in RunScheduler.

RunScheduler holds a back-reference to the facade so runLoop can dispatch
through facade.runTask via bracket access — preserves runTask as the single
per-task choreography point. StreamPump gets a back-reference to RunScheduler
via setRunScheduler() to break the import cycle.

EdgeMaterializer's two indirect calls (pushStatusFromNodeToEdges and
propagateDisabledStatus, previously routed through the facade back-ref) now
go through this.runner.runScheduler.X, restoring tighter visibility on those
methods (deleted from facade entirely).

Pure refactor: no behavior change, no public API change. 1,663 integration
tests pass unchanged.
